feat: use BaseIconButton instead of TooltipButton
